パーセプトロンを実装した。

+++ src/main/java/net/seiko_comb/combS8214808/joiss2016/LearningData.java
@@ -14,6 +14,6 @@ public class LearningData {
 	
 	@Override
 	public String toString() {
-		return String.format("[%s -> %d]", x, y);
+		return String.format("[%s -> %d]", x, (int)y);
 	}
 }
